Virtualise EM algorithm
