+++ em/em_sqdist.cpp
@@ -4,9 +4,8 @@
 #include "em_sqdist.h"
 
 EMSqdist::EMSqdist(int cluster_count_) :
-	cluster_count(cluster_count_),
-	clusters(),
-	timestamps()
+	EMCircular(cluster_count_),
+	clusters()
 {
 	double s = 0;
 	for (int i = 0; i < cluster_count; ++i) {
@@ -78,12 +77,8 @@ double EMSqdist::maximisation() {
 	return sqrt(shift);
 }
 
-double EMSqdist::time_to_phase(uint32_t time) {
-	float phase = fmodf(time, 604800.0f) / 604800;
-	if (phase > 0.5) {
-		phase -= 1;
-	}
-	return phase * M_PI * 2;
+void EMSqdist::add_time(uint32_t time) {
+	timestamps.push_back(Timestamp(time, cluster_count));
 }
 
 void EMSqdist::train() {
@@ -93,28 +88,6 @@ void EMSqdist::train() {
 		shift = maximisation();
 		std::cout << "shift = " << shift << std::endl;
 	} while (isnan(shift) || shift > 1E-3);
-}
-
-void EMSqdist::add_time(uint32_t time) {
-	timestamps.push_back(Timestamp(time, cluster_count));
-}
-
-EMSqdist::Timestamp::Timestamp(uint32_t time, int cluster_count) :
-	phase(time_to_phase(time)),
-	alpha()
-{
-	double s = 0;
-	double r;
-
-	for (int i = cluster_count; i; --i) {
-		r = float(rand()) / RAND_MAX;
-		s += r;
-		alpha.push_back(r);
-	}
-
-	for (int i = cluster_count - 1; i >= 0; --i) {
-		alpha[i] /= s;
-	}
 }
 
 EMSqdist::Cluster::Cluster() :
